T22 creazione test per ativity lista studenti iscritti
