+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "com.kasteca"
-        minSdkVersion 16
+        minSdkVersion 23
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
@@ -81,6 +81,13 @@ dependencies {
 
     implementation 'androidx.test.espresso:espresso-idling-resource:3.2.0'
     androidTestUtil 'androidx.test:orchestrator:1.1.0'
+
+    // Optional -- Hamcrest library
+    androidTestImplementation 'org.hamcrest:hamcrest-library:1.3'
+    // Optional -- UI testing with UI Automator
+    androidTestImplementation 'androidx.test.uiautomator:uiautomator:2.2.0'
+
+    androidTestImplementation 'androidx.test.espresso:espresso-intents:3.1.0'
 }
