+++ test-reports/connected/com.kasteca.activityTest.MainActivityTest.html
@@ -22,7 +22,7 @@
 <tr>
 <td>
 <div class="infoBox" id="tests">
-<div class="counter">3</div>
+<div class="counter">6</div>
 <p>tests</p>
 </div>
 </td>
@@ -34,7 +34,7 @@
 </td>
 <td>
 <div class="infoBox" id="duration">
-<div class="counter">10.822s</div>
+<div class="counter">16.334s</div>
 <p>duration</p>
 </div>
 </td>
@@ -56,6 +56,9 @@
 <li>
 <a href="#tab0">Tests</a>
 </li>
+<li>
+<a href="#tab1">Devices</a>
+</li>
 </ul>
 <div id="tab0" class="tab">
 <h2>Tests</h2>
@@ -63,27 +66,59 @@
 <thead>
 <tr>
 <th>Test</th>
-<th>Nexus_5X_API_29_x86(AVD) - 10</th>
+<th>Pixel_3_API_28(AVD) - 9</th>
+<th>Pixel_XL_API_29(AVD) - 10</th>
 </tr>
 </thead>
 <tr>
-<td>mainActivityCorrectLoginDocenteTest</td>
-<td class="success">passed (1.563s)</td>
+<td>A_mainActivityCorrectLoginStudente</td>
+<td class="success">passed (3.153s)</td>
+<td class="success">passed (3.382s)</td>
 </tr>
 <tr>
-<td>mainActivityCorrectLoginStudente</td>
-<td class="success">passed (1.696s)</td>
+<td>B_mainActivityCorrectLoginDocenteTest</td>
+<td class="success">passed (3.048s)</td>
+<td class="success">passed (2.945s)</td>
 </tr>
 <tr>
-<td>mainActivityLoginFailTest</td>
-<td class="success">passed (7.563s)</td>
+<td>C_mainActivityLoginFailTest</td>
+<td class="success">passed (1.940s)</td>
+<td class="success">passed (1.866s)</td>
+</tr>
+</table>
+</div>
+<div id="tab1" class="tab">
+<h2>Devices</h2>
+<table>
+<thead>
+<tr>
+<th>Devices</th>
+<th>Tests</th>
+<th>Failures</th>
+<th>Duration</th>
+<th>Success rate</th>
+</tr>
+</thead>
+<tr>
+<td class="success">Pixel_3_API_28(AVD) - 9</td>
+<td>3</td>
+<td>0</td>
+<td>8.141s</td>
+<td>100%</td>
+</tr>
+<tr>
+<td class="success">Pixel_XL_API_29(AVD) - 10</td>
+<td>3</td>
+<td>0</td>
+<td>8.193s</td>
+<td>100%</td>
 </tr>
 </table>
 </div>
 </div>
 <div id="footer">
 <p>Generated by 
-<a href="http://www.gradle.org">Gradle 5.6.4</a> at 9-apr-2020 11.00.33</p>
+<a href="http://www.gradle.org">Gradle 5.6.4</a> at 15-apr-2020 18.06.33</p>
 </div>
 </div>
 </body>
