+++ test-reports/connected/com.kasteca.activityTest.html
@@ -21,7 +21,7 @@
 <tr>
 <td>
 <div class="infoBox" id="tests">
-<div class="counter">3</div>
+<div class="counter">18</div>
 <p>tests</p>
 </div>
 </td>
@@ -33,7 +33,7 @@
 </td>
 <td>
 <div class="infoBox" id="duration">
-<div class="counter">10.822s</div>
+<div class="counter">1m46.39s</div>
 <p>duration</p>
 </div>
 </td>
@@ -70,11 +70,65 @@
 </thread>
 <tr>
 <td class="success">
+<a href="com.kasteca.activityTest.ListaStudentiIscrittiActivityTestConStudenti.html">ListaStudentiIscrittiActivityTestConStudenti</a>
+</td>
+<td>2</td>
+<td>0</td>
+<td>21.178s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success">
+<a href="com.kasteca.activityTest.ListaStudentiIscrittiActivityTestSenzaStudenti.html">ListaStudentiIscrittiActivityTestSenzaStudenti</a>
+</td>
+<td>2</td>
+<td>0</td>
+<td>15.459s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success">
+<a href="com.kasteca.activityTest.LoginDocenteTest.html">LoginDocenteTest</a>
+</td>
+<td>2</td>
+<td>0</td>
+<td>5.702s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success">
 <a href="com.kasteca.activityTest.MainActivityTest.html">MainActivityTest</a>
 </td>
-<td>3</td>
+<td>6</td>
 <td>0</td>
-<td>10.822s</td>
+<td>16.334s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success">
+<a href="com.kasteca.activityTest.RichiestaIscrizioneActivityTestCodiceCorsoEsatto.html">RichiestaIscrizioneActivityTestCodiceCorsoEsatto</a>
+</td>
+<td>2</td>
+<td>0</td>
+<td>19.295s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success">
+<a href="com.kasteca.activityTest.RichiestaIscrizioneActivityTestCodiceCorsoNonEsistente.html">RichiestaIscrizioneActivityTestCodiceCorsoNonEsistente</a>
+</td>
+<td>2</td>
+<td>0</td>
+<td>6.541s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success">
+<a href="com.kasteca.activityTest.RichiestaIscrizioneActivityTestRichiestaGiaInviata.html">RichiestaIscrizioneActivityTestRichiestaGiaInviata</a>
+</td>
+<td>2</td>
+<td>0</td>
+<td>21.884s</td>
 <td class="success">100%</td>
 </tr>
 </table>
@@ -82,7 +136,7 @@
 </div>
 <div id="footer">
 <p>Generated by 
-<a href="http://www.gradle.org">Gradle 5.6.4</a> at 9-apr-2020 11.00.33</p>
+<a href="http://www.gradle.org">Gradle 5.6.4</a> at 15-apr-2020 18.06.33</p>
 </div>
 </div>
 </body>
