+++ test-reports/connected/com.kasteca.html
@@ -21,7 +21,7 @@
 <tr>
 <td>
 <div class="infoBox" id="tests">
-<div class="counter">1</div>
+<div class="counter">2</div>
 <p>tests</p>
 </div>
 </td>
@@ -72,7 +72,7 @@
 <td class="success">
 <a href="com.kasteca.ExampleInstrumentedTest.html">ExampleInstrumentedTest</a>
 </td>
-<td>1</td>
+<td>2</td>
 <td>0</td>
 <td>0.001s</td>
 <td class="success">100%</td>
@@ -82,7 +82,7 @@
 </div>
 <div id="footer">
 <p>Generated by 
-<a href="http://www.gradle.org">Gradle 5.6.4</a> at 9-apr-2020 11.00.33</p>
+<a href="http://www.gradle.org">Gradle 5.6.4</a> at 15-apr-2020 18.06.33</p>
 </div>
 </div>
 </body>
