+++ test-reports/connected/index.html
@@ -19,7 +19,7 @@
 <tr>
 <td>
 <div class="infoBox" id="tests">
-<div class="counter">7</div>
+<div class="counter">20</div>
 <p>tests</p>
 </div>
 </td>
@@ -31,7 +31,7 @@
 </td>
 <td>
 <div class="infoBox" id="duration">
-<div class="counter">32.881s</div>
+<div class="counter">1m46.39s</div>
 <p>duration</p>
 </div>
 </td>
@@ -74,27 +74,18 @@
 <td class="success">
 <a href="com.kasteca.html">com.kasteca</a>
 </td>
-<td>1</td>
+<td>2</td>
 <td>0</td>
 <td>0.001s</td>
 <td class="success">100%</td>
 </tr>
 <tr>
 <td class="success">
-<a href="com.kasteca.activity.html">com.kasteca.activity</a>
-</td>
-<td>3</td>
-<td>0</td>
-<td>22.058s</td>
-<td class="success">100%</td>
-</tr>
-<tr>
-<td class="success">
 <a href="com.kasteca.activityTest.html">com.kasteca.activityTest</a>
 </td>
-<td>3</td>
+<td>18</td>
 <td>0</td>
-<td>10.822s</td>
+<td>1m46.39s</td>
 <td class="success">100%</td>
 </tr>
 </tbody>
@@ -116,25 +107,65 @@
 <tr>
 <td class="success"/>
 <a href="com.kasteca.ExampleInstrumentedTest.html">com.kasteca.ExampleInstrumentedTest</a>
-<td>1</td>
+<td>2</td>
 <td>0</td>
 <td>0.001s</td>
 <td class="success">100%</td>
 </tr>
 <tr>
 <td class="success"/>
-<a href="com.kasteca.activityTest.RichiestaIscrizioneActivityTest.html">com.kasteca.activityTest.RichiestaIscrizioneActivityTest</a>
-<td>3</td>
+<a href="com.kasteca.activityTest.ListaStudentiIscrittiActivityTestConStudenti.html">com.kasteca.activityTest.ListaStudentiIscrittiActivityTestConStudenti</a>
+<td>2</td>
 <td>0</td>
-<td>22.058s</td>
+<td>21.178s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success"/>
+<a href="com.kasteca.activityTest.ListaStudentiIscrittiActivityTestSenzaStudenti.html">com.kasteca.activityTest.ListaStudentiIscrittiActivityTestSenzaStudenti</a>
+<td>2</td>
+<td>0</td>
+<td>15.459s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success"/>
+<a href="com.kasteca.activityTest.LoginDocenteTest.html">com.kasteca.activityTest.LoginDocenteTest</a>
+<td>2</td>
+<td>0</td>
+<td>5.702s</td>
 <td class="success">100%</td>
 </tr>
 <tr>
 <td class="success"/>
 <a href="com.kasteca.activityTest.MainActivityTest.html">com.kasteca.activityTest.MainActivityTest</a>
-<td>3</td>
+<td>6</td>
 <td>0</td>
-<td>10.822s</td>
+<td>16.334s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success"/>
+<a href="com.kasteca.activityTest.RichiestaIscrizioneActivityTestCodiceCorsoEsatto.html">com.kasteca.activityTest.RichiestaIscrizioneActivityTestCodiceCorsoEsatto</a>
+<td>2</td>
+<td>0</td>
+<td>19.295s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success"/>
+<a href="com.kasteca.activityTest.RichiestaIscrizioneActivityTestCodiceCorsoNonEsistente.html">com.kasteca.activityTest.RichiestaIscrizioneActivityTestCodiceCorsoNonEsistente</a>
+<td>2</td>
+<td>0</td>
+<td>6.541s</td>
+<td class="success">100%</td>
+</tr>
+<tr>
+<td class="success"/>
+<a href="com.kasteca.activityTest.RichiestaIscrizioneActivityTestRichiestaGiaInviata.html">com.kasteca.activityTest.RichiestaIscrizioneActivityTestRichiestaGiaInviata</a>
+<td>2</td>
+<td>0</td>
+<td>21.884s</td>
 <td class="success">100%</td>
 </tr>
 </tbody>
@@ -143,7 +174,7 @@
 </div>
 <div id="footer">
 <p>Generated by 
-<a href="http://www.gradle.org">Gradle 5.6.4</a> at 9-apr-2020 11.00.33</p>
+<a href="http://www.gradle.org">Gradle 5.6.4</a> at 15-apr-2020 18.06.33</p>
 </div>
 </div>
 </body>
